Make the result stream of `Images::build` continuous (#134)
Currently, when the `Images::build` endpoint is called, no continuous
stream is returned: once an image is successfully built, a single
`ImageBuildLibpod200Response` is returned with the entire output of the
build process.

This is especially bad for user interfaces, as no feedback can be
provided to the user.

To make the stream continuous, the implementation of the endpoint is
rewritten. Here, the implementation of the endpoint `Images::pull` is
used as a template.